requirements.txt
affine==2.3.1
aiohttp==3.8.3
branca==0.5.0
earthpy==0.9.4
eo_learn_core==1.4.0
eo_learn_coregistration==1.4.0
eo_learn_features==1.4.0
eo_learn_geometry==1.4.0
eo_learn_io==1.4.0
eo_learn_mask==1.4.0
eo_learn_ml_tools==1.4.0
eo_learn_visualization==1.4.0
esda==2.4.3
folium==0.12.1.post1
geopandas==0.11.1
geopy==2.3.0
libpysal==4.6.2
matplotlib==3.1
numpy==1.23.3
oauthlib==3.2.1
pandas==1.5.0
planetary_computer==0.5.1
pystac_client==0.6.1
python_dateutil==2.8.2
rasterio==1.3.2
Requests==2.31.0
requests_oauthlib==1.3.1
scikit_fuzzy==0.4.2
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
setuptools==67.7.2
Shapely==1.8.4
splot==1.1.5.post1
utm==0.7.0
xarray==2022.12.0


